LayoutTests/fast/html/image-element-serialization-expected.txt

Test that an  element in HTML has an end tag when serialized.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS image.__proto__ is HTMLUnknownElement.prototype
PASS div.innerHTML is "<image></image>"
PASS successfullyParsed is true

TEST COMPLETE
